Finish fixing merge conflict
