+++ keras_cv/training/contrastive/contrastive_trainer.py
@@ -16,7 +16,7 @@
 import tensorflow as tf
 from tensorflow import keras
 
-from keras_cv.utils.training import convert_inputs_to_tf_dataset
+from keras_cv.utils.train import convert_inputs_to_tf_dataset
 
 
 class ContrastiveTrainer(keras.Model):
